grc/gui_qt: remove GUI connections in sync with the underlying connection
The current QT GUI "by default" operates under the assumption that the
"core" data structures are never modified or removed by someone
else (i.e. all modifications directly originate from GUI
interactions). This assumption however is sometimes false, this
situation can **sometimes** be handled by overloading methods on the
"core" classes to intercept such modifications and update the GUI in
sync, for example this is done for port clones (e.g. when changing the
number of ports on an "Add" block).

Connections are implicitly removed by the "core", for example when one
reduces the number of ports on an "Add" (or similar) block and one of
the removed ports was connected, or for example when making certain
edits to the IO signature of embedded python blocks. This connection
removal can be intercepted by overloading
`CoreFlowgraph.remove_element()` and delegating the removal to the
GUI. Note, that this overload handles all element types though I'm not
aware of any other elements besides connections which were previously
missed by the GUI.

TODO: maybe options blocks are a problem?